Add useStyle hook
