+++ packages/fuselage/src/hooks/useStyle.js
@@ -1,0 +1,6 @@
+import { useClassNameMapping } from './useClassNameMapping';
+
+export const useStyle = (cssFn, ...args) => {
+  const mapClassName = useClassNameMapping(...args);
+  return mapClassName(cssFn);
+};
